refactor: extract Node.js parsers from parsers/__init__.py into parsers/node_parsers.py
Move _npm_purl, parse_npm_packages, parse_yarn_lock, parse_pnpm_lock,
and detect_npx_package (268 lines) into a dedicated module.
parsers/__init__.py drops from 1,250 → 1,008 lines (-19%).
All symbols re-exported for backward compatibility.
